requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-Mail
email-validator
Flask-SQLAlchemy
python-dotenv
